Makefile: Fix device not booting with GCC 7.x and above
@nathanchance: Filtered from the original commit (https://github.com/zanezam/boeffla-kernel-cm-n51x0/commit/228cc26b4b66ce3c9232eb00faaf95c3b1b3ae3c)

I tested all of the flags in the above commit and this is the only
one that prevented the phone from booting, with -Os, -O2, and -O3.

Without this, the phone will boot to the Google logo, hang for a
few seconds, then reboot, and repeat the process.

This flag can be read about below. This issue can also be resolved
(supposedly) with the -mstrict-align flag, leading me to believe it
is something with memory access. I believe this is a better solution
though given that ARM handles unaligned access perfectly fine.

https://gcc.gnu.org/onlinedocs/gcc/Optimize-Options.html#index-fstore-merging
https://gcc.gnu.org/onlinedocs/gcc/AArch64-Options.html#index-mstrict-align

Change-Id: If2d6e1dc0aaae9cd073a83d351218d90a772db66
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: Harsh Shandilya <harsh@prjkt.io>
Signed-off-by: rezaadi0105 <rezaadipangestu5@gmail.com>